// src/interfaces.ts
export interface SectionInterface {
  padding?: any;
  margin?: any;
  position?: string;
  width?: string;
  minWidth?: any;
  maxWidth?: any;
  height?: string;
  maxHeight?: string;
  minHeight?: string;
  smPadding?: string;
  color?: string;
}

export interface MainHeadingInterface {
  green?: boolean;
  mt?: string;
  margin?: string;
  mb?: string;
  width?:string
}

export interface TextWrapperInterface {
  color?: string;
  size?: string;
  weight?: any;
  spacing?: string;
  padding?: string;
  margin?: string;
  mb?: string;
  mt?: string;
}

export interface ColumnInterface {
  justify?: string;
  align?: string;
  gap?: string;
  padding?: string;
  margin?: string;
  positon?: string;
  width?: string;
  minWidth?: string;
  maxWidth?: string;
  height?: string;
  maxHeight?: string;
  minHeight?: string;
  position?: string;
}

export interface EmailInterface {
  error?: boolean;
}